# Conflicts:
#	src/main/resources/http/car.http
#	src/main/webapp/app/dashboard/dashboard-api.service.ts
#	src/main/webapp/app/revenue_api/lot.service.ts

+++ src/main/java/com/example/parkinglot/entity/Price.java
@@ -11,7 +11,7 @@ import java.math.BigDecimal;
 public class Price {
 
     @Id
-    private int duration;
+    private Long duration;
 
     private BigDecimal price;
 }
